--- instrumentation/log4j/log4j-2.13.2/library-autoconfigure/src/main/resources/META-INF/services/org.apache.logging.log4j.core.util.ContextDataProvider
@@ -1,1 +1,0 @@
-io.opentelemetry.instrumentation.log4j.v2_13_2.OpenTelemetryContextDataProvider
