+++ instrumentation/log4j/log4j-2.16/javaagent/src/test/groovy/AutoLog4j2Test.groovy
