+++ instrumentation/log4j/log4j-2.16/library-autoconfigure/src/main/resources/META-INF/services/org.apache.logging.log4j.core.util.ContextDataProvider
@@ -1,0 +1,1 @@
+io.opentelemetry.instrumentation.log4j.v2_16.OpenTelemetryContextDataProvider
