+++ instrumentation/log4j/log4j-2.16/library-autoconfigure/src/test/groovy/LibraryLog4j2Test.groovy
