+++ settings.gradle.kts
@@ -265,9 +265,9 @@ include(":instrumentation:liberty:liberty:javaagent")
 include(":instrumentation:liberty:liberty-dispatcher:javaagent")
 include(":instrumentation:log4j:log4j-1.2:javaagent")
 include(":instrumentation:log4j:log4j-2.7:javaagent")
-include(":instrumentation:log4j:log4j-2.13.2:javaagent")
-include(":instrumentation:log4j:log4j-2.13.2:library")
-include(":instrumentation:log4j:log4j-2.13.2:library-autoconfigure")
+include(":instrumentation:log4j:log4j-2.16:javaagent")
+include(":instrumentation:log4j:log4j-2.16:library")
+include(":instrumentation:log4j:log4j-2.16:library-autoconfigure")
 include(":instrumentation:log4j:log4j-2-common:testing")
 include(":instrumentation:logback-1.0:javaagent")
 include(":instrumentation:logback-1.0:library")
